Generalize inertia calculation
