+++ phobos/utils/__init__.py
@@ -3,3 +3,4 @@
 # from . import transform
 # from . import tree
 # from . import urdf
+# from . import inertia
